+++ examples/transform/src/localization.ts
@@ -1,4 +1,4 @@
-import {configureTransformLocalization} from 'lit-localize';
+import {configureTransformLocalization} from '@lit/localize';
 
 export const {getLocale} = configureTransformLocalization({
   sourceLocale: 'en',